requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
DateTime==5.5
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
logging==0.4.9.6
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
pytz==2025.1
rsa==4.9
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
zope.interface==7.2
